Refactor implementation into library
