+++ lib/server.dart
@@ -1,0 +1,1 @@
+export 'package:server/src/endpoints_v1.dart';
